requirements.txt
sqlalchemy>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0
